fix: set override correctly for systemd services (#48)
Previously, I had used a priority of 500 thinking that the "default" is 1000.
Turns out that Nix sets a priority of 100 to regular assigments. Since the
`oci-containers` module sets `Restart=always`, it ends up silently overriding
any value set by compose2nix.

To fix this, I am bumping down the priority to 90. I've also added an integration
test to verify that we indeed override the NixOS module.