fix: de-duplicate env files between CLI and Compose
If the same env file is passed via CLI and set in the Compose `env_file`
for a service, then we do not want to include the env file twice.

To achieve this, we map the CLI env files into absolute paths to ensure
that we can compare them with env_file entries returned by the Compose
library.