feat: initial support for Compose Build spec (#52)
The idea is we generate a systemd (one-shot) service per build that
users can run _manually_ to build the container and add it to the runtime's
image store.

Users can also pass in `-build=true` to enable the build service to run
as a dependency for the container. However, keeping this config would result in
a _new_ build on every restart of the system (or root target), which would
implicitly result in an update of the container.

For now, we do not support Git repo build contexts (i.e., pull from repo
before building).